requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0.1
django-filter==1.1.0
django-heroku==0.2.0
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
Markdown==2.6.11
oauthlib==2.0.6
psycopg2==2.7.3.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
whitenoise==3.3.1
